Added: Chartmap as compilation unit independent from the PID loop unit
